Upgrade rubygems to 2.6.13
Case CPANEL-15622: Update rubygems to 2.6.13 to address:

CVE-2017-0902: A DNS request hijacking vulnerability.
CVE-2017-0899: An ANSI escape sequence vulnerability.
CVE-2017-0900: A DoS vulnerability in the query command.
CVE-2017-0901: A vulnerability in the gem installer that allowed
               a malicious gem to overwrite arbitrary files.

The patches were taken from the upstream notice issued in regards to
these CVEs (see case for link) - with one minor alteration to ensure
that the test suite passes on SCL builds.